+++ WPFBot3000/changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.8]
+
+ - Added RadioButton control
+ - Updated build process to update markdown docs
+
 ## [0.9.7]
 
 - Added GroupBox control